+++ Source/World/Generators/Structures_Builder.h
@@ -1,6 +1,7 @@
 #ifndef STRUCTURES_BUILDER_H_INCLUDED
 #define STRUCTURES_BUILDER_H_INCLUDED
 
+#include <vector>
 #include "../Chunk/Nodes.h"
 
 class IBlock_Accessible;
@@ -11,7 +12,7 @@ namespace Structure
     {
         struct Defer_Block
         {
-            Defer_Block(int x, int y, int z, CBlock b)
+            Defer_Block(int64_t x, int64_t y, int64_t z, CBlock b)
             :   pos     (x, y, z)
             ,   block   (b)
             { }
@@ -28,7 +29,7 @@ namespace Structure
                         CBlock fill);
 
             void makeColumn(const Block::Position& base,
-                            int height,
+                            int64_t height,
                             CBlock block);
 
             void build(IBlock_Accessible& access);
